+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 - completion for `configuration_backend`, `jwks_uri_backend`, `token_backend` and `userinfo_backend` attributes in `oidc` block [#98](https://github.com/avenga/couper-vscode/pull/98)
 - checks for invalid block labels [#98](https://github.com/avenga/couper-vscode/pull/98)
 
+### Fixed
+
+- broken icons in tooltips on vscode.dev [#101](https://github.com/avenga/couper-vscode/pull/101)
+
 ### Removed
 
 - completion for `path` attribute in `error_handler` block [#98](https://github.com/avenga/couper-vscode/pull/98)