+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - completion for `configuration_backend`, `jwks_uri_backend`, `token_backend` and `userinfo_backend` attributes in `oidc` block [#98](https://github.com/avenga/couper-vscode/pull/98)
 - checks for invalid block labels [#98](https://github.com/avenga/couper-vscode/pull/98)
 - completion for `beta_token_request` block [#100](https://github.com/avenga/couper-vscode/pull/100)
+- completion for new `grant_type`s `"password"` and `"urn:ietf:params:oauth:grant-type:jwt-bearer"` and related attributes for `oauth2` block [#105](https://github.com/avenga/couper-vscode/pull/105)
 
 ### Fixed